Extract added-method call-site guards from TransformWorkerProgram
Move the listed members out of TransformWorkerProgram without changing their bodies.
